+++ app/src/main/java/com/battlelancer/seriesguide/dataliberation/AutoBackupFragment.kt
@@ -11,9 +11,11 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.CompoundButton
 import android.widget.TextView
+import androidx.core.view.isGone
 import androidx.core.widget.TextViewCompat
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
+import androidx.lifecycle.lifecycleScope
 import com.battlelancer.seriesguide.R
 import com.battlelancer.seriesguide.SgApp
 import com.battlelancer.seriesguide.databinding.FragmentAutoBackupBinding
@@ -75,7 +77,7 @@ class AutoBackupFragment : Fragment() {
         binding.checkBoxAutoBackupCreateCopy
             .setOnCheckedChangeListener { buttonView: CompoundButton, isChecked: Boolean ->
                 BackupSettings.setCreateCopyOfAutoBackup(buttonView.context, isChecked)
-                updateFileViews()
+                viewModel.updateCopiesFileNames()
             }
 
         binding.buttonAutoBackupShowsExportFile.setOnClickListener {
@@ -97,8 +99,21 @@ class AutoBackupFragment : Fragment() {
             )
         }
 
+        viewLifecycleOwner.lifecycleScope.launch {
+            viewModel.copiesFiles.collect {
+                binding.textViewAutoBackupShowsExportFile
+                    .setFilenameOrPlaceholder(it.fileNameShows, it.placeholderText)
+                binding.textViewAutoBackupListsExportFile
+                    .setFilenameOrPlaceholder(it.fileNameLists, it.placeholderText)
+                binding.textViewAutoBackupMoviesExportFile
+                    .setFilenameOrPlaceholder(it.fileNameMovies, it.placeholderText)
+
+                binding.groupUserFiles.isGone = !it.visible
+            }
+        }
+
         binding.groupState.visibility = View.GONE
-        updateFileViews()
+        viewModel.updateCopiesFileNames()
         setProgressLock(false) // Also disables import button if backup availability unknown.
 
         // restore UI state
@@ -174,7 +189,8 @@ class AutoBackupFragment : Fragment() {
             return
         }
         viewModel.updateAvailableBackupData()
-        updateFileViews()
+        // Note: backup may remove a copy file URI
+        viewModel.updateCopiesFileNames()
         setProgressLock(false)
     }
 
@@ -204,7 +220,7 @@ class AutoBackupFragment : Fragment() {
         if (uri == null) return
         DataLiberationTools.tryToPersistUri(requireContext(), uri)
         BackupSettings.storeExportFileUri(requireContext(), type, uri, true)
-        updateFileViews()
+        viewModel.updateCopiesFileNames()
     }
 
     private fun setContainerSettingsVisible(visible: Boolean) {
@@ -227,40 +243,15 @@ class AutoBackupFragment : Fragment() {
         binding.buttonAutoBackupMoviesExportFile.isEnabled = !isLocked
     }
 
-    private fun updateFileViews() {
-        val binding = binding ?: return
-        if (BackupSettings.isCreateCopyOfAutoBackup(context)) {
-            setUriOrPlaceholder(
-                binding.textViewAutoBackupShowsExportFile,
-                BackupSettings.getExportFileUri(
-                    context, JsonExportTask.BACKUP_SHOWS, true
-                )
-            )
-            setUriOrPlaceholder(
-                binding.textViewAutoBackupListsExportFile,
-                BackupSettings.getExportFileUri(
-                    context, JsonExportTask.BACKUP_LISTS, true
-                )
-            )
-            setUriOrPlaceholder(
-                binding.textViewAutoBackupMoviesExportFile,
-                BackupSettings.getExportFileUri(
-                    context, JsonExportTask.BACKUP_MOVIES, true
-                )
-            )
-            binding.groupUserFiles.visibility = View.VISIBLE
-        } else {
-            binding.groupUserFiles.visibility = View.GONE
-        }
-    }
-
-    private fun setUriOrPlaceholder(textView: TextView, uri: Uri?) {
-        textView.text = uri?.toString() ?: getString(R.string.no_file_selected)
+    private fun TextView.setFilenameOrPlaceholder(fileName: String?, placeholder: String) {
+        text = fileName ?: placeholder
         TextViewCompat.setTextAppearance(
-            textView,
-            if (uri == null) {
+            this,
+            if (fileName == null) {
                 R.style.TextAppearance_SeriesGuide_Body2_Error
-            } else R.style.TextAppearance_SeriesGuide_Body2_Dim
+            } else {
+                R.style.TextAppearance_SeriesGuide_Body2_Dim
+            }
         )
     }