+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Releases marked with 🧪 (or previously with the "beta" suffix) were released o
 
 * 🔧 Backup: rename to "Export" and "Import", move them and Auto Backup up to More screen.
 * 🔧 Import: display names of selected files, select for import after selecting a file.
+* 🔧 Auto Backup: display names of selected files.
 
 ### 2025.2.2 - 2025-05-22 🧪